Export of internal ZetaSQL changes.
--
Change by Matthew Brown <matthewbr@google.com>:
Support setting reserved keywords in Java LanguageOptions
--
Change by Matthew Brown <matthewbr@google.com>:
ZetaSQL release formatter support in java client
--
Change by ZetaSQL Team <no-reply@google.com>:
Never add a line break before -> in a lambda expression.
--
Change by ZetaSQL Team <no-reply@google.com>:
Support inferring proto type through a subquery.
--
Change by ZetaSQL Team <no-reply@google.com>:
Throw more explicit error when a script variable is used incorrectly
--
Change by ZetaSQL Team <no-reply@google.com>:
Add a test that will fail if any language feature that isn't currently applied to a test where it shouldn't be is applied to a test where it shouldn't be.  This at least should slow down addition of cases that wrap language features around tests where they shouldn't be added.
--
Change by ZetaSQL Team <no-reply@google.com>:
Add anonymization support for TABLESAMPLE
--
Change by ZetaSQL Team <no-reply@google.com>:
Removing IdString::data()
--
Change by ZetaSQL Team <no-reply@google.com>:
Add helper functions to FunctionCallBuilder for the $case_no_value and $not functions
--
Change by ZetaSQL Team <no-reply@google.com>:
Used TypedList instead of TypedRange as a container for values of Range type
--
Change by ZetaSQL Team <no-reply@google.com>:
add helper: IntervalValue.isZero
--
Change by ZetaSQL Team <no-reply@google.com>:
Create a helper function to build the AggregateScan for the LIKE ANY and LIKE ALL <subquery> cases
--
Change by ZetaSQL Team <no-reply@google.com>:
Add ARRAY_SLICE reference implementation and compliance test
--
Change by ZetaSQL Team <no-reply@google.com>:
Update the ColumnFactory::MakeCol function to support collation when creating columns
--
Change by ZetaSQL Team <no-reply@google.com>:
Changes to the semantics of the LIKE ALL operator to comply with discussed changes when handling NULL patterns
--
Change by ZetaSQL Team <no-reply@google.com>:
fix bug that COLLATE() function validates number of arguments in a wrong position.
--
Change by ZetaSQL Team <no-reply@google.com>:
Add a KitchenSinkEnumPB type with a required enum to the test_schema
--
Change by ZetaSQL Team <no-reply@google.com>:
Delete a pointer to Type from TypedList since it's not used
--
Change by ZetaSQL Team <no-reply@google.com>:
Rollback @Nullable annotation added in CL 464181150.
--
Change by ZetaSQL Team <no-reply@google.com>:
Add non-determinism to evaluation context for ARRAY_FILTER and ARRAY_TRANSFORM reference implementation function eval
--
Change by ZetaSQL Team <no-reply@google.com>:
Fix returning clause sql builder when there are table alias and column alias.
--
Change by ZetaSQL Team <no-reply@google.com>:
Fix bug in reference implementation of $like_any_array and $like_all_array causing errors for certain ARRAY cases
--
Change by ZetaSQL Team <no-reply@google.com>:
Add error messages to opaque `Preconditions.checkState()` calls in `Value`, so callers can tell what mistake they made without reading the code.
--
Change by ZetaSQL Team <no-reply@google.com>:
Support "RANGE" type in parser
--
Change by ZetaSQL Team <no-reply@google.com>:
When the parse location record type is FULL_SCOPE, make the location of an array element call include both the array expr and the index, like what the parser AST does.
--
Change by ZetaSQL Team <no-reply@google.com>:
Add ARRAY_SLICE reference implementation and compliance test
--
Change by ZetaSQL Team <no-reply@google.com>:
Documentation for WITH expression
--
Change by ZetaSQL Team <no-reply@google.com>:
Add ARRAY_SLICE reference implementation and compliance test
--
Change by ZetaSQL Team <no-reply@google.com>:
Implement range type serialization/deserialization to/from TypeProto.
--
Change by ZetaSQL Team <no-reply@google.com>:
Fix the default set of rewrites for the execute_query tool to be the 'ALL' set, instead of 'NONE'.
--
Change by ZetaSQL Team <no-reply@google.com>:
Fix regex to match string literals with ')' in parser tests
--
Change by ZetaSQL Team <no-reply@google.com>:
Add ARRAY_SLICE builtin function signature and rewriter
--
Change by ZetaSQL Team <no-reply@google.com>:
Fix NOLINT comment - START/END should be matching
--
Change by ZetaSQL Team <no-reply@google.com>:
Fix collation propagation on resolved column for rewriter's internal API AnalyzeSubstitute by re-designing the substitute framework.
--
Change by ZetaSQL Team <no-reply@google.com>:
Rename the test_outcome field of ComplianceTestCaseLables to test_error_mode so it's more clear when reading a message of this time what that field means.
--
Change by ZetaSQL Team <no-reply@google.com>:
fix bug that the variables inside a procedure are checked against the variables outside of the procedure.
--
Change by ZetaSQL Team <no-reply@google.com>:
Make ValidateTypeProto error message more clear if passed argument is empty proto
--
Change by ZetaSQL Team <no-reply@google.com>:
Resolve bug in FlattenRewriter that generated invalid ResolvedASTs for some query shapes.
--
Change by ZetaSQL Team <no-reply@google.com>:
Accommodate NULL templates in RET_CHECK() post-resolving Format() calls.
--
Change by ZetaSQL Team <no-reply@google.com>:
Minor change to FORMAT error handling for invalid specifiers.
--
Change by ZetaSQL Team <no-reply@google.com>:
remove the flag to defer folding errors to runtime.
--
Change by Matthew Brown <matthewbr@google.com>:
Fix parse tree for IN expression
--
Change by ZetaSQL Team <no-reply@google.com>:
Factor code for processing the file based test segments out of SQLTestBase.
Provides a strongly typed API for SQLTestBase to interact with the [key=...] options in filebase test segments.
More importantly, this consolidates the validating and parsing of those options into a single place instead of being spread all over the expansive SQLTestBase class. In some cases (such as required_features) the same option was being parsed as each time it was used.
--
Change by Matthew Brown <matthewbr@google.com>:
ZetaSQL: Fix RoundingMode in java client
--
Change by ZetaSQL Team <no-reply@google.com>:
Canonicalize CAST-ing and FORMAT-ing NaN and 0 to not be signed.
--
Change by ZetaSQL Team <no-reply@google.com>:
Add compliance test labels to undeclared output file
--
Change by Matthew Brown <matthewbr@google.com>:
ZetaSQL: Fix formatter for ZetaSQL, release basic test
--
Change by Matthew Brown <matthewbr@google.com>:
Fix parse tree for BETWEEN and << expressions
  (And 289 more changes)

GitOrigin-RevId: 58a2a39ffafcb968074d8d3217b186df3a532287

+++ zetasql/public/civil_time.cc
@@ -386,7 +386,9 @@ int64_t DatetimeValue::Packed64DatetimeSeconds() const {
 }
 
 int64_t DatetimeValue::Packed64DatetimeMicros() const {
-  return (Packed64DatetimeSeconds() << kMicrosShift) | (nanosecond_ / 1000);
+  return static_cast<int64_t>(
+      (static_cast<uint64_t>(Packed64DatetimeSeconds()) << kMicrosShift) |
+      (nanosecond_ / 1000));
 }
 
 std::string DatetimeValue::DebugString() const {
